Move the HybridRecordReceiver into the Sierra transformer
