+++ catalogue_pipeline/transformer/transformer_sierra/src/main/scala/uk/ac/wellcome/platform/transformer/sierra/services/SierraTransformerWorkerService.scala
@@ -5,7 +5,6 @@ import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.SQSStream
 import uk.ac.wellcome.models.transformable.SierraTransformable
-import uk.ac.wellcome.platform.transformer.receive.HybridRecordReceiver
 import uk.ac.wellcome.platform.transformer.sierra.SierraTransformableTransformer
 
 import scala.concurrent.Future
